Text.ChordPro.Types

Description

This module defines types used for working with ChordPro data.

We break this into four main types: * a Lyric relates to what the specification calls a "song line." * a Directive is either a preamble, formatting, or output-related directive. * a ChordProLine is a sum containing one of the above two types * a ChordProSong is a list of ChordProLines

Documentation

data Lyric

A Lyric relates to a song line. Given a line in a ChordPro file, we will run across both lyrics and chords.

If we have the input string Swing [D]low, sweet [G]chari[D]ot, then we can parse that into [Lyric "Swing ", Chord D, Lyric "low, sweet ", Chord G, Lyric "chari", Chord D, Lyric "ot"].

Constructors

Lyric String

Chord String